Alot of changes...
It's been a while since I committed so it kind of piled up
To keep it all compilable It's for the best to just commit it all at
once

Moved from Random to SecureRandom using RNGCryptoServiceProvider
Removed "GetNextRandom"X from the SSPClient, it will no longer ask the
server for a random number
Removed messages MsgGetNextId, MsgGetNextIdResponse
Added CreateNewThread, it will create a new thread for "onConnect" when
the haskshake is finished
MazingHandshake will no longer throw a error if you had a bad seed while
solving the maze in-memory
OperationalSockets have a Packet Queue now so they run more smoothly
while multiple are connected
Fixed the handshake system server sided it's non-blocking now
SimpleRsaHandshake will now provide a Dual-Fingerprint, providing MD5
and SHA512
A few fixed for the XML Hiding Layer

+++ SecureSocketProtocol3/Security/Encryptions/Compiler/InstructionInfo.cs
@@ -160,7 +160,7 @@ namespace SecureSocketProtocol3.Security.Encryptions.Compiler
             get
             {
                 if (_value2_long == 0)
-                    _value2_long = (byte)Value2.IntValue();
+                    _value2_long = (byte)Value2.LongValue();
                 return _value2_long;
             }
         }
@@ -170,7 +170,7 @@ namespace SecureSocketProtocol3.Security.Encryptions.Compiler
             get
             {
                 if (_value3_long == 0)
-                    _value3_long = (byte)Value3.IntValue();
+                    _value3_long = (byte)Value3.LongValue();
                 return _value3_long;
             }
         }
